Bug fixes and type safety
Moved to unsigned ints for position and color, and liberal use of const
for data points and clusters. Added an outline to the datapoint visual
marker. Moved to cluster counter inside the loops rather than vector
subtraction to find the index. Trying out random location placement of a
cluster without any associated data points, but there is a bug in the
assignment step so it just keeps randomly jumping around the plane.